+++ TuProfe/app/src/main/java/com/example/tuprofe/ui/profe/ProfeState.kt
@@ -6,6 +6,6 @@ import com.example.tuprofe.data.ReviewInfo
 data class ProfeState(
     val profesor: Profesor? = null,
     val professorReviews: List<ReviewInfo> = emptyList(),
-    val averageRating: Int = 0,
+    val averageRating: Float = 0f,
     val isLoading: Boolean = true
 )
