+++ TuProfe/app/src/main/java/com/example/tuprofe/ui/search/SearchState.kt
@@ -5,7 +5,7 @@ import com.example.tuprofe.data.Profesor
 data class SearchState(
     val searchQuery: String = "",
     val searchResults: List<Profesor> = emptyList(),
-    val ratings: Map<String, Int> = emptyMap(),
+    val ratings: Map<String, Float> = emptyMap(),
     val isLoading: Boolean = false,
     val navigateToProfileId: Int? = null
 )
